+++ package.json
@@ -45,6 +45,7 @@
     "@types/node": "^25.0.3",
     "@types/ws": "^8.18.1",
     "babel-jest": "^30.2.0",
+    "esbuild": "^0.27.2",
     "jest": "^30.2.0",
     "nodemon": "^3.1.10",
     "prettier": "^3.7.4",
